Tests/LibWeb/Text/expected/credential-protection-exfiltration-detection.txt

=== Credential Protection: Exfiltration Detection Test ===
undefined
Test 1: Form element analysis
  form exists: true
  form.action: https://evil.example.com/steal
  form.method: POST
undefined
Test 2: Credential field detection
  password field exists: true
  password field type: password
  password field has value: true
undefined
Test 3: Cross-origin analysis
  current origin: file://
  current host: localhost
  form action URL: https://evil.example.com/steal
  action host: evil.example.com
  is cross-origin: true
undefined
Test 4: Suspicious indicators detected
  has password field: true
  cross-origin submission: true
  action to external domain: true
  suspicious domain pattern: true
undefined
Test 5: Expected FormMonitor checks
  - Should detect password field: YES
  - Should detect cross-origin submission: YES
  - Should trigger credential exfiltration alert: YES
  - Alert type: CREDENTIAL_EXFILTRATION
  - Severity: HIGH or CRITICAL
undefined
Test 6: URL pattern analysis
  protocol: https:
  hostname: evil.example.com
  pathname: /steal
  is absolute URL: true
  points to different domain: true
undefined
=== Expected Behavior ===
For a cross-origin credential submission:
  1. FormMonitor detects password field
  2. FormMonitor detects form action points to different origin
  3. Security alert is triggered via on_credential_exfiltration_alert
  4. SecurityAlertDialog is shown to user
  5. Form submission is blocked until user decision
undefined
=== Alert Details ===
Expected alert properties:
  alert_type: credential_exfiltration or third_party_form_post
  severity: high or critical
  form_origin: file://
  action_origin: https://evil.example.com
  uses_https: true
  has_password_field: true
  is_cross_origin: true
undefined
=== Test Complete ===
This is a credential exfiltration scenario and SHOULD trigger alert.
